+++ packages/frontend/src/components/InlineLabel.tsx
@@ -1,7 +1,9 @@
+import { Tooltip } from '@mui/material';
 import { classNames } from '../utils';
 
 interface InlineLabelProps {
   text: string;
+  title?: string;
   button?: JSX.Element;
   className?: string;
   onClick?: () => void;
@@ -9,21 +11,24 @@ interface InlineLabelProps {
 
 export const InlineLabel = ({
   text,
+  title,
   button,
   className,
   onClick,
 }: InlineLabelProps): JSX.Element => {
   return (
-    <span
-      onClick={onClick}
-      className={classNames(
-        className,
-        'h-6 pl-1 pr-1 mr-1 text-xs text-white no-underline bg-gray-800 py-[1px] rounded',
-        onClick ? 'cursor-pointer' : ''
-      )}
-    >
-      {text}
-      {button && button}
-    </span>
+    <Tooltip placement="top-start" title={!title ? '' : title} arrow>
+      <span
+        onClick={onClick}
+        className={classNames(
+          className,
+          'h-6 pl-1 pr-1 mr-1 text-xs text-white no-underline bg-gray-800 py-[1px] rounded',
+          onClick ? 'cursor-pointer' : ''
+        )}
+      >
+        {text}
+        {button && button}
+      </span>
+    </Tooltip>
   );
 };