+++ packages/frontend/src/navigation/AuthenticatedRoutes.tsx
@@ -1,12 +1,10 @@
-import { StartupLoader } from '../startupLoader';
-import NotFoundPage from '@/pages/NotFoundPage';
 import SettingsPage from '@/pages/Settings/SettingsPage';
 import StartPage from '@/pages/Start/StartPage';
-import { ROLE_ADMIN, ROLE_QUANTIFIER, ActiveUserRoles } from '@/model/auth';
+import { ActiveUserRoles, ROLE_ADMIN, ROLE_QUANTIFIER } from '@/model/auth';
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
+import { difference } from 'lodash';
 import { useRecoilValue } from 'recoil';
-import AuthenticatedLayout from '../layouts/AuthenticatedLayout';
 
 const MyPraisePage = React.lazy(() => import('@/pages/MyPraise/MyPraisePage'));
 const UsersPage = React.lazy(() => import('@/pages/Users/UsersPage'));
@@ -33,36 +31,42 @@ const QuantifyPage = React.lazy(
   () => import('@/pages/QuantifyPeriodReceiver/QuantifyPeriodReceiverPage')
 );
 
+const EventLogsPage = React.lazy(
+  () => import('@/pages/EventLogs/EventLogsPage')
+);
+
+const FAQPage = React.lazy(() => import('@/pages/FAQ/FAQPage'));
+
 interface AuthRouteProps {
   children: JSX.Element;
-  exact?: boolean;
+  userRoles: string[];
   path: string;
-  roles: string[];
+  exact?: boolean;
+  roles?: string[];
 }
 // A Route that takes an array of roles as argument and redirects
 // to frontpage if user do not belong to any of the given roles
-const AuthRoute = ({ children, ...props }: AuthRouteProps): JSX.Element => {
-  const userRoles = useRecoilValue(ActiveUserRoles);
-
-  let authenticated = false;
-  for (const role of props.roles) {
-    if (userRoles?.includes(role)) {
-      authenticated = true;
-      break;
-    }
-  }
+const AuthRoute = ({
+  children,
+  userRoles,
+  path,
+  exact = false,
+  roles = [],
+}: AuthRouteProps): JSX.Element => {
+  // Check if user.roles contain all required roles
+  const hasRequiredRoles = difference(roles, userRoles).length === 0;
 
   return (
     <Route
-      {...props}
-      render={({ location }): JSX.Element =>
-        authenticated ? (
+      path={path}
+      exact={exact}
+      render={(): JSX.Element =>
+        hasRequiredRoles ? (
           children
         ) : (
           <Redirect
             to={{
-              pathname: '/',
-              state: { from: location },
+              pathname: '/404',
             }}
           />
         )
@@ -71,18 +75,24 @@ const AuthRoute = ({ children, ...props }: AuthRouteProps): JSX.Element => {
   );
 };
 
-const Routes = (): JSX.Element => {
+const AuthenticatedRoutes = (): JSX.Element | null => {
+  const userRoles = useRecoilValue(ActiveUserRoles);
+
   return (
     <Switch>
       <Route path="/mypraise">
         <MyPraisePage />
       </Route>
 
-      <AuthRoute roles={[ROLE_ADMIN]} path={'/pool'}>
+      <AuthRoute userRoles={userRoles} roles={[ROLE_ADMIN]} path={'/pool'}>
         <UsersPage />
       </AuthRoute>
 
-      <AuthRoute roles={[ROLE_ADMIN]} path={'/periods/createupdate'}>
+      <AuthRoute
+        userRoles={userRoles}
+        roles={[ROLE_ADMIN]}
+        path={'/periods/createupdate'}
+      >
         <PeriodsCreateUpdatePage />
       </AuthRoute>
 
@@ -90,14 +100,23 @@ const Routes = (): JSX.Element => {
         <PeriodReceiverSummaryPage />
       </Route>
 
+      <Route exact path="/eventlogs">
+        <EventLogsPage />
+      </Route>
+
       <AuthRoute
+        userRoles={userRoles}
         roles={[ROLE_QUANTIFIER]}
         path={'/periods/:periodId/quantify/receiver/:receiverId'}
       >
         <QuantifyPage />
       </AuthRoute>
 
-      <AuthRoute roles={[ROLE_QUANTIFIER]} path={'/periods/:periodId/quantify'}>
+      <AuthRoute
+        userRoles={userRoles}
+        roles={[ROLE_QUANTIFIER]}
+        path={'/periods/:periodId/quantify'}
+      >
         <QuantifyPeriodPage />
       </AuthRoute>
 
@@ -113,7 +132,7 @@ const Routes = (): JSX.Element => {
         <PraiseDetailsPage />
       </Route>
 
-      <AuthRoute roles={[ROLE_ADMIN]} path={'/settings'}>
+      <AuthRoute userRoles={userRoles} roles={[ROLE_ADMIN]} path={'/settings'}>
         <SettingsPage />
       </AuthRoute>
 
@@ -121,21 +140,18 @@ const Routes = (): JSX.Element => {
         <StartPage />
       </Route>
 
+      <Route exact path="/faq">
+        <FAQPage />
+      </Route>
+
       <Route path="/*">
-        <NotFoundPage />
+        <Redirect
+          to={{
+            pathname: '/404',
+          }}
+        />
       </Route>
     </Switch>
-  );
-};
-
-const AuthenticatedRoutes = (): JSX.Element => {
-  return (
-    <>
-      <StartupLoader />
-      <AuthenticatedLayout>
-        <Routes />
-      </AuthenticatedLayout>
-    </>
   );
 };